Implement action for saving items
