--- src/utils/save_item.js
@@ -1,37 +1,0 @@
-"use strict";
-
-function onStart() {
-  window.EditorsNotes.events.emit('loadstart');
-}
-
-function onStop(ret) {
-  window.EditorsNotes.events.emit('loadstop');
-  return ret;
-}
-
-function onError(error) {
-  window.EditorsNotes.events.emit('loadstop');
-  throw error;
-}
-
-
-
-module.exports = function (type, id, projectURL, data) {
-  var apiFetch = require('./api_fetch')
-    , isNew = id === null
-    , method = isNew ? 'post' : 'put'
-    , url = `${projectURL}${type}s/`
-
-  if (!isNew) url += `${id}/`;
-
-  if (type === 'topic' && !isNew) {
-    url += 'w/';
-  }
-
-  onStart();
-
-  return (
-    apiFetch(url, { method, body: JSON.stringify(data) })
-      .then(onStop, onError)
-  )
-}
